Rename crate to `indexed_docs`
