+++ crates/assistant/src/assistant.rs
@@ -20,9 +20,9 @@ pub(crate) use completion_provider::*;
 pub(crate) use context_store::*;
 use fs::Fs;
 use gpui::{actions, AppContext, Global, SharedString, UpdateGlobal};
+use indexed_docs::IndexedDocsStore;
 pub(crate) use inline_assistant::*;
 pub(crate) use model_selector::*;
-use rustdoc::IndexedDocsStore;
 use semantic_index::{CloudEmbeddingProvider, SemanticIndex};
 use serde::{Deserialize, Serialize};
 use settings::{Settings, SettingsStore};
