+++ crates/indexed_docs/src/indexer.rs
