+++ crates/indexed_docs/src/store.rs
